requirements.txt
# Core data science libraries
pandas
numpy<2
matplotlib
seaborn
scikit-learn
# Deep learning frameworks
torch==2.2.0
torchvision==0.17.0
# Image processing
Pillow
# Web development
django==5.0.1
python-dotenv # for loading .env files
django-bootstrap5

# Loading and utilities
tqdm
